*x64 build for windows (exe and vanila game dll) *add nvml lib and gpu fan speed (win64 only)
git-svn-id: https://svn.code.sf.net/p/quake2xp/code@1520 d83f5309-02f3-4a42-a5f4-12e6af65da40

+++ trunk/client/cg_screen.c
@@ -564,10 +564,10 @@ int entitycmpfnc (const entity_t * a, const entity_t * b) {
 	 ** all other models are sorted by model then skin
 	 */
 	if (a->model == b->model) {
-		return ((int)a->skin - (int)b->skin);
+		return ((intptr_t)a->skin - (intptr_t)b->skin);
 	}
 	else {
-		return ((int)a->model - (int)b->model);
+		return ((intptr_t)a->model - (intptr_t)b->model);
 	}
 }
