+++ trunk/win32/win_network.c
@@ -65,7 +65,7 @@ void NetadrToSockadr (netadr_t *a, struct sockaddr *s) {
 	}
 	else if (a->type == NA_IP) {
 		((struct sockaddr_in *)s)->sin_family = AF_INET;
-		((struct sockaddr_in *)s)->sin_addr.s_addr = *(int *)&a->ip;
+		((struct sockaddr_in *)s)->sin_addr.s_addr = *(intptr_t *)&a->ip;
 		((struct sockaddr_in *)s)->sin_port = a->port;
 	}
 /*	else if (a->type == NA_IPX) {
@@ -85,7 +85,7 @@ void NetadrToSockadr (netadr_t *a, struct sockaddr *s) {
 void SockadrToNetadr (struct sockaddr *s, netadr_t *a) {
 	if (s->sa_family == AF_INET) {
 		a->type = NA_IP;
-		*(int *)&a->ip = ((struct sockaddr_in *)s)->sin_addr.s_addr;
+		*(intptr_t *)&a->ip = ((struct sockaddr_in *)s)->sin_addr.s_addr;
 		a->port = ((struct sockaddr_in *)s)->sin_port;
 	}
 /*	else if (s->sa_family == AF_IPX) {
